refactor(redux): refactor reducer `entitiesForAuthors`
